Leeds

+++ src/Services/Gene/Gene.Application/EventHandlers/GeneEventHandler_Core.cs
@@ -1,4 +1,5 @@
 
+using AutoMapper;
 using CQRS.Core.Exceptions;
 using Daikon.Events.Gene;
 using Gene.Application.Contracts.Persistence;
@@ -11,20 +12,15 @@ namespace Gene.Application.Query.EventHandlers
         private readonly IGeneRepository _geneRepository;
         private readonly IGeneEssentialityRepository _geneEssentialityRepository;
         private readonly IGeneProteinProductionRepository _geneProteinProductionRepository;
-
         private readonly IGeneProteinActivityAssayRepository _geneProteinActivityAssayRepository;
-
         private readonly IGeneHypomorphRepository _geneHypomorphRepository;
-
         private readonly IGeneCrispriStrainRepository _geneCrispriStrainRepository;
-
         private readonly IGeneResistanceMutationRepository _geneResistanceMutationRepository;
-
         private readonly IGeneVulnerabilityRepository _geneVulnerabilityRepository;
-
         private readonly IGeneUnpublishedStructuralInformationRepository _geneUnpublishedStructuralInformationRepository;
-
+        private readonly IGeneExpansionPropRepo _geneExpansionPropRepo;
         private readonly ILogger<GeneEventHandler> _logger;
+        private readonly IMapper _mapper;
 
         public GeneEventHandler(IGeneRepository geneRepository, 
                                 IGeneEssentialityRepository geneEssentialityRepository,
@@ -35,6 +31,8 @@ namespace Gene.Application.Query.EventHandlers
                                 IGeneResistanceMutationRepository geneResistanceMutationRepository,
                                 IGeneVulnerabilityRepository geneVulnerabilityRepository,
                                 IGeneUnpublishedStructuralInformationRepository geneUnpublishedStructuralInformationRepository,
+                                IGeneExpansionPropRepo geneExpansionPropRepo,
+                                IMapper mapper,
                                 ILogger<GeneEventHandler> logger)
         {
             _geneRepository = geneRepository ?? throw new ArgumentNullException(nameof(geneRepository));
@@ -46,28 +44,17 @@ namespace Gene.Application.Query.EventHandlers
             _geneResistanceMutationRepository = geneResistanceMutationRepository ?? throw new ArgumentNullException(nameof(geneResistanceMutationRepository));
             _geneVulnerabilityRepository = geneVulnerabilityRepository ?? throw new ArgumentNullException(nameof(geneVulnerabilityRepository));
             _geneUnpublishedStructuralInformationRepository = geneUnpublishedStructuralInformationRepository ?? throw new ArgumentNullException(nameof(geneUnpublishedStructuralInformationRepository));
+            _geneExpansionPropRepo = geneExpansionPropRepo ?? throw new ArgumentNullException(nameof(geneExpansionPropRepo));
+            _mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         }
 
         public async Task OnEvent(GeneCreatedEvent @event)
         {
             _logger.LogInformation("OnEvent: GeneCreatedEvent: {Id}", @event.Id);
-            var gene = new Domain.Entities.Gene
-            {
-                Id = @event.Id,
-                StrainId = @event.StrainId,
-                Name = @event.Name,
-
-                AccessionNumber = @event.AccessionNumber,
-                Function = @event.Function,
-                Product = @event.Product,
-                FunctionalCategory = @event.FunctionalCategory,
-
-                DateCreated = DateTime.UtcNow,
-                IsModified = false,
-                IsDraft = false
-            };
-
+            var gene = _mapper.Map<Domain.Entities.Gene>(@event);
+            gene.Id = @event.Id;
+            
             try
             {
                 await _geneRepository.CreateGene(gene);
@@ -82,15 +69,19 @@ namespace Gene.Application.Query.EventHandlers
         public async Task OnEvent(GeneUpdatedEvent @event)
         {
             _logger.LogInformation("OnEvent: GeneUpdatedEvent: {Id}", @event.Id);
-            var gene = await _geneRepository.ReadGeneById(@event.Id);
+            var existingGene = await _geneRepository.ReadGeneById(@event.Id);
 
-            gene.StrainId = @event.StrainId;
-            gene.Name = @event.Name;
-            gene.AccessionNumber = @event.AccessionNumber;
-            gene.Function = @event.Function;
-            gene.Product = @event.Product;
-            gene.FunctionalCategory = @event.FunctionalCategory;
-            gene.IsModified = true;
+            if (existingGene == null)
+            {
+                throw new EventHandlerException(nameof(EventHandler), $"GeneUpdatedEvent Error updating gene with id @event.Id", new Exception("Gene not found"));
+            }
+
+            var gene = _mapper.Map<Domain.Entities.Gene>(existingGene);
+            _mapper.Map(@event, gene);
+            
+            // Preserve the original creation date and creator
+            gene.DateCreated = existingGene.DateCreated;
+            gene.CreatedById = existingGene.CreatedById;
 
             try
             {
@@ -109,6 +100,11 @@ namespace Gene.Application.Query.EventHandlers
             _logger.LogInformation("OnEvent: GeneDeletedEvent: {Id}", @event.Id);
             var gene = await _geneRepository.ReadGeneById(@event.Id);
 
+            if (gene == null)
+            {
+                throw new EventHandlerException(nameof(EventHandler), $"GeneDeletedEvent Error deleting gene with id @event.Id", new Exception("Gene not found"));
+            }
+
             try
             {
                 await _geneEssentialityRepository.DeleteAllEssentialitiesOfGene(gene.Id);
@@ -119,6 +115,7 @@ namespace Gene.Application.Query.EventHandlers
                 await _geneResistanceMutationRepository.DeleteAllResistanceMutationsOfGene(gene.Id);
                 await _geneVulnerabilityRepository.DeleteAllVulnerabilitiesOfGene(gene.Id);
                 await _geneUnpublishedStructuralInformationRepository.DeleteAllUnpublishedStructuralInformationsOfGene(gene.Id);
+                await _geneExpansionPropRepo.DeleteAllOfEntity(gene.Id);
                 await _geneRepository.DeleteGene(gene.Id);
             }
             catch (RepositoryException ex)
@@ -127,5 +124,6 @@ namespace Gene.Application.Query.EventHandlers
             }
 
         }
+
     }
 }
